Add some function
changed:      ft_atoi.c
new file:     ft_memalloc.c
new file:     ft_memdel.c
new file:     ft_strclr.c
new file:     ft_strdel.c
new file:     ft_strnew.c
changed:      libft.h
changed:      makefile

+++ ft_atoi.c
@@ -24,7 +24,7 @@ static int	ft_isspace(int c)
 int			ft_atoi(const char *nptr)
 {
 	int			sign;
-	long long	result;
+	int	result;
 
 	sign = 1;
 	result = 0;
